pyproject.toml
[tool.poetry]
name = "auto-trade"
version = "0.1.0"
description = ""
authors = ["sandropark <birdob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
requests = "^2.31.0"
pytz = "^2024.1"
pyupbit = "^0.2.33"
flask = "^3.0.2"
gspread = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
